requirements.txt
flask==1.1.1
pillow==6.1.0
selenium==3.141.0